Fix bug with suggested search
There was a bug wherein if you typed a query, such as "bat", you'd get the
queries for "bat" but if you then  backspaced the query to "ba", the queries for
"bat" remained. Fixed this behaviour so that the query results disappear if you
reduce query length below the threshold.